🚐 Move sorting function to helper file
